pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "icedrive-authentication"
version = "0.0.1"
authors = [
  { name="Antonio Gómez ", email="antonio.gomez21@alu.uclm.es" },
]
description = "Implementation of IceDrive Authentication service"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Operating System :: OS Independent",
]
license = "GPL-3.0-only"
dependencies = [
    "zeroc-ice>=3.7",
]

[project.optional-dependencies]
tests = ["pytest"]
linters = ["black", "ruff", "pydocstyle"]

[project.urls]
"Homepage" = "https://github.com/YOURUSERNAME/YOURREPOSITORY"

[project.scripts]
# icedrive-authentication = "icedrive_authentication.app:main"
# icedrive-client = "icedrive_authentication.client:main"
icedrive-authentication="icedrive_authentication.command_line_handlers:server"
icedrive-client=" icedrive_authentication.command_line_handlers:client"